# Use multiple threads to generate object index cache.