--- a/src/openrct2/Diagnostic.cpp
+++ b/src/openrct2/Diagnostic.cpp
@@ -16,6 +16,7 @@
 
 #include <cstdarg>
 #include <cstdio>
+#include "core/String.hpp"
 #include "Diagnostic.h"
 
 #ifdef __ANDROID__
@@ -80,46 +81,45 @@ static constexpr const char * _level_strings[] = {
 void diagnostic_log(DIAGNOSTIC_LEVEL diagnosticLevel, const char *format, ...)
 {
     va_list args;
+    if (_log_levels[diagnosticLevel])
+    {
+        // Level
+        auto prefix = String::StdFormat("%s: ", _level_strings[diagnosticLevel]);
 
-    if (!_log_levels[diagnosticLevel])
-        return;
+        // Message
+        va_start(args, format);
+        auto msg = String::StdFormat_VA(format, args);
+        va_end(args);
 
-    FILE * stream = diagnostic_get_stream(diagnosticLevel);
-
-    // Level
-    fprintf(stream, "%s: ", _level_strings[diagnosticLevel]);
-
-    // Message
-    va_start(args, format);
-    vfprintf(stream, format, args);
-    va_end(args);
-
-    // Line terminator
-    fprintf(stream, "\n");
+        auto stream = diagnostic_get_stream(diagnosticLevel);
+        fprintf(stream, "%s%s\n", prefix.c_str(), msg.c_str());
+    }
 }
 
 void diagnostic_log_with_location(DIAGNOSTIC_LEVEL diagnosticLevel, const char *file, const char *function, sint32 line, const char *format, ...)
 {
     va_list args;
+    if (_log_levels[diagnosticLevel])
+    {
+        // Level and source code information
+        std::string prefix;
+        if (_log_location_enabled)
+        {
+            prefix = String::StdFormat("%s[%s:%d (%s)]: ", _level_strings[diagnosticLevel], file, line, function);
+        }
+        else
+        {
+            prefix = String::StdFormat("%s: ", _level_strings[diagnosticLevel]);
+        }
 
-    if (!_log_levels[diagnosticLevel])
-        return;
+        // Message
+        va_start(args, format);
+        auto msg = String::StdFormat_VA(format, args);
+        va_end(args);
 
-    FILE * stream = diagnostic_get_stream(diagnosticLevel);
-
-    // Level and source code information
-    if (_log_location_enabled)
-        fprintf(stream, "%s[%s:%d (%s)]: ", _level_strings[diagnosticLevel], file, line, function);
-    else
-        fprintf(stream, "%s: ", _level_strings[diagnosticLevel]);
-
-    // Message
-    va_start(args, format);
-    vfprintf(stream, format, args);
-    va_end(args);
-
-    // Line terminator
-    fprintf(stream, "\n");
+        auto stream = diagnostic_get_stream(diagnosticLevel);
+        fprintf(stream, "%s%s\n", prefix.c_str(), msg.c_str());
+    }
 }
 
 #endif // __ANDROID__