--- a/src/openrct2/core/FileIndex.hpp
+++ b/src/openrct2/core/FileIndex.hpp
@@ -20,8 +20,6 @@
 #include <string>
 #include <tuple>
 #include <vector>
-#include <thread>
-#include <mutex>
 #include "../common.h"
 #include "Console.hpp"
 #include "File.h"