--- a/src/openrct2/core/JobPool.hpp
+++ b/src/openrct2/core/JobPool.hpp
@@ -20,6 +20,7 @@
 #include <deque>
 #include <functional>
 #include <atomic>
+#include <vector>
 
 class JobPool
 {