… OpenRCT2.xcodeproj/project.pbxproj
… distribution/changelog.txt
… src/openrct2/core/String.cpp
… src/openrct2/core/String.hpp
… test/testpaint/CMakeLists.txt