This leaves quite a remarkable performance boost on my end, the results:
```
Release:
	before:
	Finished building object index in 11.13 seconds.
	after:
	Finished building object index in 0.99 seconds.
	
Debug:
	before:
	Finished building object index in 31.25 seconds.
	after:
	Finished building object index in 20.43 seconds.
```