@janisozaur 2018-04-22
1.46s -> 0.5s with warm cache for me in a quick test, but I'd still like to take a look at the diff. That'll unfortunately have to wait at least until tomorrow.