@Broxzier 2018-04-22
Very nice to see that much of an improvement! It goes from ~43 to only ~12 seconds on my machine (HDD).

One way that I think will improve this further and will allow printing the progress, is by using a job pool (or thread pool). You'll have one main thread that will build a queue of jobs to do (scanning one or multiple objects), and multiple worker threads that will perform them. The workers can already be spawned while the job queue is being filled. Through an `std::condition_variable` the main thread notifies the worker threads about new jobs for them to perform. The first thread that sees it will perform it - if they're all busy with a job already, the first job to finishes its current job will take it. Once the job queue is built, the main thread will become mostly idle, meaning that core will become available for another worker threads, so this will work on single-core CPUs too.

With this design you won't have issues where one thread can take longer to scan its range than the others, and it will allow the main thread to print out the progress (best done when a worker treads notifies the main thread that it's done with a job and take the next from the job queue, which is protected by a mutex).