@ZehMatt 2018-04-23
Makes sense, however so far this is the only parallel use case so I didn't think of making a generic system currently has much use. But perhaps we could feed it the other work as well for scenarios and track design indexing.