@IntelOrca 2018-04-23
> But perhaps we could feed it the other work as well for scenarios and track design indexing.

You have already changed the generic implementation used by all 3 indexers, no extra work is required.