@ZehMatt 2018-04-23
Oh yeah I see, didn't pay attention on who is using FileIndex.