@IntelOrca 2018-04-24
JobQueue should just be a single hpp file, no cpp file. `.hpp` should only be used for header files which are also source files and have no other corresponding source files.