@ZehMatt 2018-04-24
I wanted at some point to have dedicated worker threads so not each JobPool would spawn new ones. I can move it into the hpp for now.