@Broxzier 2018-04-24
When there are many objects with an image table that's longer than expected, the warning gets printed to the console which is not thread safe, possibly resulting in strange output in the console:

```
[ITMSAVBC] Warning: Image table size longer than expected.
[KNGBRPA2] Warning: Image table size longer than expected.
[KLOTRCA3] Warning: Image table size longer than expected.
[KNG-DRCF] Warning: Image table size longer than expected.[KNGCBST1] Warning: Image table size longer than expected.

[KNGCBST2] Warning: Image table size longer than expected.
[KNGDBRP2] Warning: Image table size longer than expected.
```

The same thing can probably happen with the progress report.

I also noticed that the progress report starts counting from ~2800/3234 objects for me, rather than from 0.