@ZehMatt 2018-04-24
The progress starting at 2800~ is because thats how many it has processed by the time it calls join() since the threads spawned already. Regarding the output order, well without having a lock on all prints I'm not sure how one would solve this.