@ZehMatt 2018-04-24
After reading all the reviews and comments I'll do one more commit tackling those last remaining things.