@ZehMatt 2018-04-24
This should for this PR, the concurrency issue with printing needs to be solved in another PR and @janisozaur seems to have some plans already.