@spacek531 2018-04-24
building objects.idx has improved from 14 seconds to 8 seconds with "warm cache" on windows