@AaronVanGeffen 2018-04-26
Per the last commit, this fails to build on macOS. While `std::string_view` may be covered, it is clear Xcode 9.3 is far from C++17-ready, unfortunately:

```
❌  /Users/travis/build/OpenRCT2/OpenRCT2/src/openrct2/object/../core/FileIndex.hpp:244:21: cannot form a reference to 'void'
                auto& items = containers.emplace_back();
```